Add frenchVeracity method and accumulate NLP modules
