--- Jupyter/text.txt
@@ -1,23 +1,0 @@
-Apple abandons AirPower wireless charging product
-Dave Lee
-North America technology reporter
-In a highly unusual step for the firm, Apple has given up on a product because it could not make it work adequately.
-
-AirPower, announced in 2017, was a mat meant to charge multiple devices without needing to plug them in.
-
-But it is understood the firm's engineers were perhaps unable to stop the mat from getting too hot.
-
-"After much effort, we've concluded AirPower will not achieve our high standards and we have cancelled the project," the firm said.
-
-The company did not elaborate further.
-
-However, rumours of issues with the product had been circulating since its announcement in September 2017. At the time, Apple said it would be released to the public some time in 2018, promising a "world-class wireless charging solution".
-
-But late last year, with the product absent from the most recent iPhone launch, noted Apple insider Jon Gruber wrote: "There are engineers who looked at AirPower's design and said it could never work, thermally, and now those same engineers have that 'told you so' smug look on their faces."
-
-In a statement emailed to the BBC, Dan Riccio, Apple's head of hardware engineering, said: "We apologise to those customers who were looking forward to this launch. We continue to believe that the future is wireless and are committed to push the wireless experience forward."
-
-The cancelled product could affect sales of the company's wireless headphones - AirPods - which were promoted and sold with the promise they would be charged using AirPower in future. Packaging for the AirPods contained a diagram of how AirPower would work.
-
-Apple's rivals, such as Huawei and Samsung, have already released products that charge their devices wirelessly.
-
