--- dataBase/amir.json
@@ -1,13 +1,0 @@
-{
-  "1": {
-    "Name": "Amir",
-    "Family": "SHIRALI POUR",
-    "Cell Phone": "0628299022"
-  },
-  "2": {
-    "Name": "Imane",
-    "Family": "Lamriou",
-    "Cell Phone": "0767456609*",
-    "class": "21254465"
-  }
-}
