--- dataBase/claims_uycheck.csv
@@ -1,11 +1,0 @@
-;claims;links;date_publish;source
-0;Explicador: esto es lo que debe chequear en la papeleta antes de votar para el balotaje;http://uycheck.com/explicador-esto-es-lo-que-debe-chequear-en-la-papeleta-antes-de-votar-para-el-balotaje/;noviembre 24, 2019;Nacionales 2019
-1;Debate | Daniel Martínez: “Bajamos la desocupación a poco más de 5%”;http://uycheck.com/__trashed/;noviembre 21, 2019;Debate
-2;Debate | Daniel Martinez: “El Estado creció sus funcionarios en educación, vivienda y salud, y decreció en todas el resto de las áreas”;http://uycheck.com/daniel-martinez-el-estado-crecio-sus-funcionarios-en-educacion-vivienda-y-salud-y-decrecio-en-todas-el-resto-de-las-areas/;noviembre 21, 2019;Debate
-3;Debate | Lacalle Pou: [Licitaciones públicas observadas] “en las intendencias del FA, el año 2018, 62%. Las del PN la mitad”;http://uycheck.com/lacalle-pou-las-licitaciones-publicas-observadas-en-las-intendencias-las-del-fa-el-ano-2018-62-las-del-pn-la-mitad/;noviembre 21, 2019;Debate
-4;Es falso que en 2005 el desempleo haya sido del 20%, como afirmó Martínez;http://uycheck.com/es-falso-que-en-2005-el-desempleo-haya-sido-del-20-como-afirmo-martinez/;noviembre 14, 2019;Debate
-5;Es falso que hoy sólo 27% de los niños y 8,8% de las embarazadas estén correctamente controlados en ASSE, como afirmó Lacalle Pou;http://uycheck.com/es-falso-que-hoy-solo-27-de-los-ninos-y-88-de-las-embarazadas-esten-correctamente-controlados-en-asse-como-afirmo-lacalle-pou/;noviembre 14, 2019;Debate
-6;Jorge Gandini: “70 mil personas es la diferencia entre que se aprobara y no se aprobara”;http://uycheck.com/jorge-gandini-70-mil-personas-es-la-diferencia-entre-que-se-aprobara-y-no-se-aprobara/;noviembre 08, 2019;Cortita y al pie
-7;Enrique Rubio: “En Rivera, donde hay el 1% de las rapiñas del país, [que] el 60% apoye la reforma es un dato”;http://uycheck.com/enrique-rubio-en-rivera-donde-hay-el-1-de-las-rapinas-del-pais-que-el-60-apoye-la-reforma-es-un-dato/;noviembre 08, 2019;Frente Amplio
-8;Luis Lacalle Pou: “Desde el primer gobierno de Sanguinetti a la fecha, [el Uruguay] ha decidido reducir las FFAA”;http://uycheck.com/luis-lacalle-pou-desde-el-primer-gobierno-de-sanguinetti-a-la-fecha-el-uruguay-ha-decidido-reducir-las-ffaa/;noviembre 07, 2019;Defensa
-9;Julio María Sanguinetti: “En plena crisis del 2002 (…) había 5.500 rapiñas, o algo así. Hoy hay 30.000”;http://uycheck.com/julio-maria-sanguinetti-en-plena-crisis-del-2002-habia-5-500-rapinas-o-algo-asi-hoy-hay-30-000/;noviembre 06, 2019;Partido Colorado
