+++ requirements.txt
@@ -1,8 +1,11 @@
 appnope==0.1.0
 attrs==19.3.0
+autopep8==1.4.4
 backcall==0.1.0
-bleach==3.1.4
-certifi==2019.11.28
+beautifulsoup4==4.8.2
+bleach==3.1.1
+bs4==0.0.1
+certifi==2020.4.5.1
 cffi==1.14.0
 chardet==3.0.4
 control==0.8.3
@@ -13,14 +16,18 @@ cytoolz==0.9.0.1
 decorator==4.4.2
 defusedxml==0.6.0
 dill==0.2.9
+en-core-web-sm==2.0.0
 entrypoints==0.3
+flake8==3.7.9
 future==0.18.2
 idna==2.9
 importlib-metadata==1.5.0
 ipykernel==5.1.4
+ipynb==0.5.1
 ipython==7.13.0
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
+it-core-news-sm==2.0.0
 jedi==0.16.0
 Jinja2==2.11.1
 jsonschema==3.2.0
@@ -29,8 +36,10 @@ jupyter-client==6.0.0
 jupyter-console==6.1.0
 jupyter-core==4.6.3
 kiwisolver==1.1.0
+lxml==4.5.0
 MarkupSafe==1.1.1
 matplotlib==3.1.3
+mccabe==0.6.1
 mistune==0.8.4
 msgpack==0.6.1
 msgpack-numpy==0.4.3.2
@@ -40,6 +49,7 @@ nbformat==5.0.4
 nltk==3.4.5
 notebook==6.0.3
 numpy==1.18.1
+pandas==1.0.3
 pandocfilters==1.4.2
 parso==0.6.2
 pexpect==4.8.0
@@ -49,21 +59,26 @@ preshed==2.0.1
 prometheus-client==0.7.1
 prompt-toolkit==3.0.4
 ptyprocess==0.6.0
+pycodestyle==2.5.0
 pycparser==2.20
+pyflakes==2.1.1
 Pygments==2.6.1
 pyOpenSSL==19.1.0
 pyparsing==2.4.6
 pyrsistent==0.15.7
 PySocks==1.7.1
 python-dateutil==2.8.1
+pytz==2020.1
 pyzmq==19.0.0
 qtconsole==4.7.1
 QtPy==1.9.0
-regex==2020.2.20
+regex==2020.4.4
 requests==2.23.0
 scipy==1.4.1
 Send2Trash==1.5.0
 six==1.14.0
+slycot==0.3.5.0
+soupsieve==2.0
 spacy==2.0.16
 tagme==0.1.3
 terminado==0.8.3
